修改参数接受类型

+++ src/main/java/so/sao/shop/supplier/service/QualificationService.java
@@ -17,7 +17,7 @@ public interface QualificationService {
      * @param reason
      * @return
      */
-    Result updateQualificationStatus(Integer accountId, Integer qualificationStatus, String reason );
+    Result updateQualificationStatus(Long accountId, Integer qualificationStatus, String reason );
 
     /**
      *查询登录供应商的资质状态
@@ -66,5 +66,5 @@ public interface QualificationService {
      * @param accountId
      * @return Result
      */
-    Result updateQualificationMessageRead( Integer accountId );
+    Result updateQualificationMessageRead( Long accountId );
 }
